# Speed up calibration