Fixes #89

After this PR, you can pass `--mc_dropout_num 0` to `genienlp predict ` and then  `--fast` to `genienlp calibrate` and get faster calibrators.

Results on a batch of 50 examples on CPU:
No calibrators:
- 6 seconds

One calibrator:
- Before this PR, with `--mc_dropout_num 5`: 44 seconds
- After removing unused confidence features, with  `--mc_dropout_num 5`: 32 seconds
- With `--mc_dropout_num 0` and `--fast` : 8 seconds

And our metric AUCPR (area under the precision-recall curve) on the almond dataset is worse by just ~0.5%.